Python/requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.2
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
bcrypt==4.0.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
channels-redis==4.0.0
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
crcmod==1.7
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.8
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-comment-migrate==0.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-ranged-response==0.2.0
django-redis==5.2.0
django-simple-captcha==0.5.17
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
django-sslserver==0.22
dnspython==2.2.1
drf-yasg==1.21.4
eventlet==0.33.1
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
jpush==3.3.9
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.4
mysqlclient==2.1.1
oss2==2.18.2
packaging==21.3
paramiko==2.11.0
Pillow==9.2.0
prompt-toolkit==3.0.31
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.15.0
PyJWT==2.5.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.4
redis==4.3.4
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tencentcloud-sdk-python==3.0.745
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.4
ua-parser==0.16.1
uritemplate==4.1.1
urllib3==1.26.12
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
uwsgi==2.0.22
